Time channels file hidden
Nobody uses TCB files anymore so I have deleted the controls.
I have also fixed a problem with the Time Unit not updating.
The P, D, S, W file selecter did not work if set to read-only, so I
have switched it back. This should be investigated in the future.