Added entries' info label, adjusted button names
Entry result information is displayed in place of the old "Page number:" label. The result details look like "Entries 26-50 of 64", so the user knows where they currently are and how many entries are in total.

Button names changed from Next to Prev (ambiguous) to "Older" and "Newer" (relating to entry run number, is more clear).