Updated results counting method
The count method will now take into account the search filters (WHERE) instead of just counting the whole table each time.
Created new method in JournalSearch to be used in JournalModel.

Binded the value of lblResults to JournalViewModel's resultsInfo, although currently the value is not updated on the first run, but is on the 2nd (so it is always one step behind).
For now the value of lblResults is directly written in JournalViewerView, although this needs to be changed to conform with the other labels.