Removed old count method, set fixed width for lblResults
Removed old SQL query counting method in the model as that is now handled by the new one in JournalSearch.

Fixed width won't break positioning of other elements in case of wrong label text value, which is also stored in the tooltip in case it is too long.

Would like to change fixed width to something more fluid, but not sure what.

Another option would be having the label below the controls, but this makes the interface look awkward and takes up a whole row for just a few words.