Fixed maxPage not updating on instrument switch
When switching to an instrument with no journal entries, the "maxPage" value would become 0 as a result of "(int) Math.ceil(totalResults / (double) PAGE_SIZE)". The spinner would then be given the value 0 as its maximum, lower than its minimum of 1, so the widget would ignore the command.

To fix this issue, "calcTotalPages" was changed to return 1 (maximum pages) instead of 0 when there are no entries.