Removed spinner, adjusted page nav methods, added text field for page number (just display for now)
Changed page control methods to return the next page value to be used by the setPageNumber method, instead of calling it directly from inside. This also makes it possible to keep the progress bar indicator since newer/olderPageNumber() are passed in as an argument for setPageNumber().

Text field only displays the current page number at the moment. Will have to make it possible to change pages by typing the page number inside it, but for that filtering out any non-number characters is required, as well as adding a listener on text change (as SWT.Selection doesn't seem to work?)