Fixed whitespace double-click error, replaced loops with streams, added full-row selection
An error was triggered when double clicking on an empty space in table when no item was selected, as okPressed() didn't check if there even is a selection before pressing the ok button (which should be disabled in this case so the user is unable to press it, but it can still activate when double clicking due to the mouse listener).

Added a check for the okPressed() override in MultipleConfigsSelectionDialog to see if the currently selected item is in the available list.
RecentConfigSelectionDialog was already doing this to avoid the issue, but the other dialogs weren't.