Added columns parameter to createTable, ConfigInfo unit tests
Removed duplicate createTable code from RecentConfigSelectionDialog.
Added parameter to the SelectionDialog createTable to take a column names list. This way, synoptics won't have two columns as only the name is needed, and the other dialogs can have two (or more, if need arises in the future).

Added unit tests for ConfigInfo names/descriptions methods, although the '...withoutCurrent' variants are using a Singleton which makes it difficult to test, and is not worth it considering the method's simplicity.