Fixed page number not displaying sometimes, resized elements, adjusted text field navigation
Page number would not be displayed due to bindValue adding a comma when 1000 and over (e.g. 1,000). Added an update strategy to the binding value to convert the source property from integer to string before updating the target.

Expanded label width size to fit results info up to "999999-999999 of 999999).
Expanded page number text field width size to fit up to page "99999".
Renamed "Refresh data" button to "Refresh" in order to make space for these changes.

Adjusted page number text field navigation so that when a number is entered, if it is out of bounds due to it being too high, it will be automatically changed to the maximum page number instead. (e.g. Max page no: 7500, Input: 99999, set page: 7500)