Added tests for synoptic components icons and types
Added tests to check if all used icons (height) and thumbnails conform to the image size standards, as well as their type (needs to be .png).
Also added test to check if there are any currently used types that are not declared in the ComponentType class.

Added two more methods in ComponentIcons to retrieve the filename a type corresponds to as String, rather than an Image object.

Todo: change print statements with proper logging.