Fixed descriptions not matching names, removed casts, other minor changes
Replaced the Lists of names and descriptions with a SortedMap that has both, to fix descriptions not matching names.

Removed all casts as they may cause problems in the future, and changed return type of ConfigInfo methods to List instead of Collection.

Added new ConfigInfo methods for SortedMap.

Added mismatched lists check for recent configs (names and timestamps).