Added types to ComponentType, resized icons/thumbs, changed some OPI types
Added DEBUG, (SANS2D) VACUUM_TANK and HE_LEVEL_GAUGE types to the ComponentType.

Changed opi_info.xml GALIL_ENGINEERING and UNCONFIRMED to DEBUG.

Resized icons and thumbnails to conform to conventions.

Changed IconConventionsTest to IconConventionsCheck (as it is more of a system test than a unit test).
Added IconConventionsCheckTest to unit test the conventions checks.

Added more checks to see if a declared type has an icon, and if each icon has a corresponding thumbnail and vice versa.